Charger basculté dans le controller
