+++ app/src/main/java/fr/haronman/demineur/fx/JeuFX.java
@@ -1,14 +1,11 @@
 package fr.haronman.demineur.fx;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.Optional;
 
 import fr.haronman.demineur.Sauvegarde;
 import fr.haronman.demineur.controller.jeu.MenuController;
 import fr.haronman.demineur.model.Difficulte;
 import fr.haronman.demineur.model.Jeu;
-import fr.haronman.demineur.model.Partie;
 import fr.haronman.demineur.model.Plateau.Case.Case;
 import fr.haronman.demineur.model.Plateau.Case.Mine;
 import fr.haronman.demineur.model.Plateau.Case.Terrain;
@@ -47,13 +44,11 @@ import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.text.Font;
-import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import javafx.util.Duration;
 
 public class JeuFX {
     private Jeu jeu;
-    private MenuBar menuBarre;
     private VBox plateauFX;
     private Stage stage;
     private Label nbrDrapeauFX;
@@ -66,7 +61,6 @@ public class JeuFX {
     public JeuFX(Stage stage, Jeu jeu){
         this.jeu = jeu;
         this.stage = stage;
-        this.menuBarre = new MenuBar();
         this.nbrDrapeauFX = new Label();
         nbrDrapeauFX.setFont(Font.font(20));
         this.visage = new ImageView(Visage.IDLE.getImage());
@@ -81,11 +75,92 @@ public class JeuFX {
     }
 
     public void show(VBox partie) throws Exception{
-
         updateTimerFX();
         StackPane root = new StackPane();
 
-        createMenuBarre();
+        MenuBar menuBarre = new MenuBar();
+
+        Menu jeuBarre = new Menu("Jeu");
+
+        MenuItem sauvegarder = new MenuItem("Sauvegarder");
+        sauvegarder.setId("save");
+        sauvegarder.setAccelerator(new KeyCodeCombination(KeyCode.S, KeyCombination.CONTROL_DOWN));
+        if(jeu.getFin()){
+            sauvegarder.setDisable(true);
+        }
+        sauvegarder.setOnAction(new MenuController(this));
+
+        MenuItem charger = new MenuItem("Charger");
+        charger.setId("load");
+        charger.setAccelerator(new KeyCodeCombination(KeyCode.L, KeyCombination.CONTROL_DOWN));
+        charger.setOnAction(new MenuController(this));
+
+        MenuItem setPause = new MenuItem("Pause / Reprendre");
+        setPause.setId("pause");
+        setPause.setAccelerator(new KeyCodeCombination(KeyCode.P, KeyCombination.CONTROL_DOWN));
+        setPause.setOnAction(new MenuController(this));
+
+        SeparatorMenuItem ligne = new SeparatorMenuItem();
+
+        MenuItem tableauScore = new MenuItem("Tableau des scores");
+        tableauScore.setId("tabScore");
+        tableauScore.setAccelerator(new KeyCodeCombination(KeyCode.T, KeyCombination.CONTROL_DOWN, KeyCombination.ALT_DOWN));
+        tableauScore.setOnAction(action -> {
+            TableauScoreFX.show();
+            System.err.println("Tableau des scores non implémenté");
+        });
+
+        MenuItem quitter = new MenuItem("Quitter");
+        quitter.setAccelerator(new KeyCodeCombination(KeyCode.Q, KeyCombination.CONTROL_DOWN));
+        quitter.setOnAction(action -> {
+            stage.close();
+        });
+
+        // Choix difficulté
+        Menu choix_difficulte = new Menu("Difficulté");
+        ToggleGroup tg = new ToggleGroup();
+        RadioMenuItem facile = new RadioMenuItem("Facile"); facile.setId("0");
+        RadioMenuItem intermediaire = new RadioMenuItem("Intermediaire"); intermediaire.setId("1");
+        RadioMenuItem difficile = new RadioMenuItem("Difficile"); difficile.setId("2");
+        RadioMenuItem expert = new RadioMenuItem("Expert"); expert.setId("3");
+        RadioMenuItem impossible = new RadioMenuItem("Impossible"); impossible.setId("4");
+        RadioMenuItem hardcore = new RadioMenuItem("Hardcore"); hardcore.setId("5");
+        RadioMenuItem diabolique = new RadioMenuItem("Diabolique"); diabolique.setId("6");
+        facile.setToggleGroup(tg);
+        intermediaire.setToggleGroup(tg);
+        difficile.setToggleGroup(tg);
+        expert.setToggleGroup(tg);
+        impossible.setToggleGroup(tg);
+        hardcore.setToggleGroup(tg);
+        diabolique.setToggleGroup(tg);
+        tg.getToggles().get(jeu.getPartie().getDifficulte().getId()).setSelected(true);
+
+        tg.selectedToggleProperty().addListener(
+            (obs, oldT, newT) -> {
+                if(newT != null){
+                    RadioMenuItem rmi = (RadioMenuItem) newT;
+                    Optional<Difficulte> d = Difficulte.getById(Integer.valueOf(rmi.getId()));
+                    if(d.isPresent()){
+                        try {
+                            updateVisage(Visage.IDLE);
+                            chrono.stop();
+                            jeu.start(d.get());
+                        } catch (Exception e) {
+                            e.printStackTrace();
+                        }
+                    }
+                }
+            }
+        );
+
+        Menu plus = new Menu("Plus");
+        MenuItem regles = new MenuItem("Règles");
+        
+        jeuBarre.getItems().addAll(sauvegarder, charger, setPause, ligne, tableauScore, quitter);
+        choix_difficulte.getItems().addAll(facile, intermediaire, difficile, expert, impossible, hardcore, diabolique);
+        plus.getItems().addAll(regles);
+
+        menuBarre.getMenus().addAll(jeuBarre, choix_difficulte, plus);
 
         root.getChildren().add(menuBarre);
         StackPane.setAlignment(menuBarre, Pos.TOP_CENTER);
@@ -500,120 +575,6 @@ public class JeuFX {
         chrono.setCycleCount(Timeline.INDEFINITE);
     }
 
-    public void createMenuBarre(){
-        Menu jeuBarre = new Menu("Jeu");
-
-        MenuItem sauvegarder = new MenuItem("Sauvegarder");
-        sauvegarder.setId("save");
-        sauvegarder.setAccelerator(new KeyCodeCombination(KeyCode.S, KeyCombination.CONTROL_DOWN));
-        if(jeu.getFin()){
-            sauvegarder.setDisable(true);
-        }
-        sauvegarder.setOnAction(new MenuController(this));
-
-        MenuItem charger = new MenuItem("Charger");
-        charger.setId("load");
-        charger.setAccelerator(new KeyCodeCombination(KeyCode.L, KeyCombination.CONTROL_DOWN));
-        charger.setOnAction(action -> {
-            try {
-                charger();
-            } catch (ClassNotFoundException | IOException e) {
-                e.printStackTrace();
-            }
-        });
-
-        MenuItem setPause = new MenuItem("Pause / Reprendre");
-        setPause.setId("pause");
-        setPause.setAccelerator(new KeyCodeCombination(KeyCode.P, KeyCombination.CONTROL_DOWN));
-        setPause.setOnAction(new MenuController(this));
-
-        SeparatorMenuItem ligne = new SeparatorMenuItem();
-
-        MenuItem tableauScore = new MenuItem("Tableau des scores");
-        tableauScore.setId("tabScore");
-        tableauScore.setAccelerator(new KeyCodeCombination(KeyCode.T, KeyCombination.CONTROL_DOWN, KeyCombination.ALT_DOWN));
-        tableauScore.setOnAction(action -> {
-            TableauScoreFX.show();
-            System.err.println("Tableau des scores non implémenté");
-        });
-
-        MenuItem quitter = new MenuItem("Quitter");
-        quitter.setAccelerator(new KeyCodeCombination(KeyCode.Q, KeyCombination.CONTROL_DOWN));
-        quitter.setOnAction(action -> {
-            stage.close();
-        });
-
-        // Choix difficulté
-        Menu choix_difficulte = new Menu("Difficulté");
-        ToggleGroup tg = new ToggleGroup();
-        RadioMenuItem facile = new RadioMenuItem("Facile"); facile.setId("0");
-        RadioMenuItem intermediaire = new RadioMenuItem("Intermediaire"); intermediaire.setId("1");
-        RadioMenuItem difficile = new RadioMenuItem("Difficile"); difficile.setId("2");
-        RadioMenuItem expert = new RadioMenuItem("Expert"); expert.setId("3");
-        RadioMenuItem impossible = new RadioMenuItem("Impossible"); impossible.setId("4");
-        RadioMenuItem hardcore = new RadioMenuItem("Hardcore"); hardcore.setId("5");
-        RadioMenuItem diabolique = new RadioMenuItem("Diabolique"); diabolique.setId("6");
-        facile.setToggleGroup(tg);
-        intermediaire.setToggleGroup(tg);
-        difficile.setToggleGroup(tg);
-        expert.setToggleGroup(tg);
-        impossible.setToggleGroup(tg);
-        hardcore.setToggleGroup(tg);
-        diabolique.setToggleGroup(tg);
-        tg.getToggles().get(jeu.getPartie().getDifficulte().getId()).setSelected(true);
-
-        tg.selectedToggleProperty().addListener(
-            (obs, oldT, newT) -> {
-                if(newT != null){
-                    RadioMenuItem rmi = (RadioMenuItem) newT;
-                    Optional<Difficulte> d = Difficulte.getById(Integer.valueOf(rmi.getId()));
-                    if(d.isPresent()){
-                        try {
-                            updateVisage(Visage.IDLE);
-                            chrono.stop();
-                            jeu.start(d.get());
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
-                    }
-                }
-            }
-        );
-
-        Menu plus = new Menu("Plus");
-        MenuItem regles = new MenuItem("Règles");
-        
-        jeuBarre.getItems().addAll(sauvegarder, charger, setPause, ligne, tableauScore, quitter);
-        choix_difficulte.getItems().addAll(facile, intermediaire, difficile, expert, impossible, hardcore, diabolique);
-        plus.getItems().addAll(regles);
-
-        menuBarre.getMenus().addAll(jeuBarre, choix_difficulte, plus);
-    }
-
-    private void charger() throws ClassNotFoundException, IOException{
-        chrono.stop();
-        FileChooser fc = new FileChooser();
-        fc.setTitle("Choississez une sauvegarde");
-        //Voir uniquement les fichiers .save
-        FileChooser.ExtensionFilter extFilter = new FileChooser.ExtensionFilter("Sauvegardes", "*.save");
-        fc.getExtensionFilters().add(extFilter);
-        fc.setInitialDirectory(new File(Sauvegarde.CHEMIN_SAUVEGARDE));
-
-        File sauvegarde = fc.showOpenDialog(stage);
-        if(sauvegarde != null){
-            Optional<Partie> loadPartie = jeu.load(sauvegarde);
-            loadPartie.ifPresent(p -> {
-                try {
-                    jeu.start(p);
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            });
-        }else{
-            chrono.play();
-        }
-    }
-
     public Jeu getJeu() {
         return jeu;
     }
@@ -624,10 +585,6 @@ public class JeuFX {
 
     public Label getNbrDrapeauFX() {
         return nbrDrapeauFX;
-    }
-
-    public MenuBar getMenuBarre() {
-        return menuBarre;
     }
 
     public VBox getPlateauFX() {
